pyproject.toml
[project]
name = "nhi-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.115.0",
    "langchain-text-splitters>=1.1.2",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.35.0",
]
